Fix a few bugs in the code and got first project deployed on device. There are some issues, if you tap button too quickly it crashes and I am having a hard time measuring the signal on my very cheap oscilloscope so may need to change code some to more effectively measure signal. But it does appear to be toggling from what I can see on the scope.

+++ Matrix/app/src/main/java/com/borderlessdesignsga/matrix/CarrierWave.java
@@ -4,6 +4,8 @@ import android.media.AudioFormat;
 import android.media.AudioManager;
 import android.media.AudioTrack;
 
+import java.util.Arrays;
+
 /**
  * Created by Wes on 9/11/2014.
  */
@@ -57,9 +59,9 @@ public class CarrierWave {
             int start = i * length;
             
             if(i % 2 == 0) //even
-                Arrays.fill(buffer, start, start + length, 0);
+                Arrays.fill(buffer, start, start + length, (byte)0);
             else //odd
-                Arrays.fill(buffer, start, start + length, 127);
+                Arrays.fill(buffer, start, start + length, (byte)127);
         }
         
         /*
